Split verifyPostDelivery into own file
