+++ packages/shared/src/store/postActions/index.ts
@@ -1,1 +1,3 @@
+export * from './logger';
 export * from './postActions';
+export * from './verifyPostDelivery';
