+++ packages/shared/src/store/postActions/logger.ts
@@ -1,0 +1,3 @@
+import { createDevLogger } from '../../debug';
+
+export const logger = createDevLogger('postActions', false);
